Crate javascriptcore_sys

Structs

JSClassDefinition

JSStaticFunction

JSStaticValue

OpaqueJSClass

OpaqueJSContext

OpaqueJSContextGroup

OpaqueJSPropertyNameAccumulator

OpaqueJSPropertyNameArray

OpaqueJSString

OpaqueJSValue

Enums

JSType

JSTypedArrayType

Functions

JSCheckScriptSyntax
Checks for syntax errors in a string of JavaScript.

JSClassCreate
Creates a JavaScript class suitable for use with JSObjectMake.

JSClassRelease
Releases a JavaScript class.

JSClassRetain
Retains a JavaScript class.

JSContextGetGlobalContext
Gets the global context of a JavaScript execution context.

JSContextGetGlobalObject
Gets the global object of a JavaScript execution context.

JSContextGetGroup
Gets the context group to which a JavaScript execution context belongs.

JSContextGroupCreate
Creates a JavaScript context group.

JSContextGroupRelease
Releases a JavaScript context group.

JSContextGroupRetain
Retains a JavaScript context group.

JSEvaluateScript
Evaluates a string of JavaScript.

JSGarbageCollect
Performs a JavaScript garbage collection.

JSGlobalContextCopyName
Gets a copy of the name of a context.

JSGlobalContextCreate
Creates a global JavaScript execution context.

JSGlobalContextCreateInGroup
Creates a global JavaScript execution context in the context group provided.

JSGlobalContextRelease
Releases a global JavaScript execution context.

JSGlobalContextRetain
Retains a global JavaScript execution context.

JSGlobalContextSetName
Sets the remote debugging name for a context.

JSObjectCallAsConstructor

JSObjectCallAsFunction

JSObjectCopyPropertyNames

JSObjectDeleteProperty

JSObjectGetArrayBufferByteLength

JSObjectGetArrayBufferBytesPtr

JSObjectGetPrivate

JSObjectGetProperty

JSObjectGetPropertyAtIndex

JSObjectGetPrototype

JSObjectGetTypedArrayBuffer

JSObjectGetTypedArrayByteLength

JSObjectGetTypedArrayByteOffset

JSObjectGetTypedArrayBytesPtr

JSObjectGetTypedArrayLength

JSObjectHasProperty

JSObjectIsConstructor

JSObjectIsFunction

JSObjectMake

JSObjectMakeArray

JSObjectMakeArrayBufferWithBytesNoCopy

JSObjectMakeConstructor

JSObjectMakeDate

JSObjectMakeError

JSObjectMakeFunction

JSObjectMakeFunctionWithCallback

JSObjectMakeRegExp

JSObjectMakeTypedArray

JSObjectMakeTypedArrayWithArrayBuffer

JSObjectMakeTypedArrayWithArrayBufferAndOffset

JSObjectMakeTypedArrayWithBytesNoCopy

JSObjectSetPrivate

JSObjectSetProperty

JSObjectSetPropertyAtIndex

JSObjectSetPrototype

JSPropertyNameAccumulatorAddName

JSPropertyNameArrayGetCount

JSPropertyNameArrayGetNameAtIndex

JSPropertyNameArrayRelease

JSPropertyNameArrayRetain

JSStringCreateWithCharacters
Creates a JavaScript string from a buffer of Unicode characters.

JSStringCreateWithUTF8CString
Creates a JavaScript string from a null-terminated UTF8 string.

JSStringGetCharactersPtr
Returns a pointer to the Unicode character buffer that serves as the backing store for a JavaScript string.

JSStringGetLength
Returns the number of Unicode characters in a JavaScript string.

JSStringGetMaximumUTF8CStringSize
Returns the maximum number of bytes a JavaScript string will take up if converted into a null-terminated UTF8 string.

JSStringGetUTF8CString
Converts a JavaScript string into a null-terminated UTF8 string, and copies the result into an external byte buffer.

JSStringIsEqual
Tests whether two JavaScript strings match.

JSStringIsEqualToUTF8CString
Tests whether a JavaScript string matches a null-terminated UTF8 string.

JSStringRelease
Releases a JavaScript string.

JSStringRetain
Retains a JavaScript string.

JSValueCreateJSONString
Creates a JavaScript string containing the JSON serialized representation of a JS value.

JSValueGetType
Returns a JavaScript value's type.

JSValueGetTypedArrayType
Returns a JavaScript value's Typed Array type.

JSValueIsArray
Tests whether a JavaScript value is an array.

JSValueIsBoolean
Tests whether a JavaScript value's type is the boolean type.

JSValueIsDate
Tests whether a JavaScript value is a date.

JSValueIsEqual
Tests whether two JavaScript values are equal, as compared by the JS == operator.

JSValueIsInstanceOfConstructor
Tests whether a JavaScript value is an object constructed by a given constructor, as compared by the JS instanceof operator.

JSValueIsNull
Tests whether a JavaScript value's type is the null type.

JSValueIsNumber
Tests whether a JavaScript value's type is the number type.

JSValueIsObject
Tests whether a JavaScript value's type is the object type.

JSValueIsObjectOfClass
Tests whether a JavaScript value is an object with a given class in its class chain.

JSValueIsStrictEqual
Tests whether two JavaScript values are strict equal, as compared by the JS === operator.

JSValueIsString
Tests whether a JavaScript value's type is the string type.

JSValueIsUndefined
Tests whether a JavaScript value's type is the undefined type.

JSValueMakeBoolean
Creates a JavaScript value of the boolean type.

JSValueMakeFromJSONString
Creates a JavaScript value from a JSON formatted string.

JSValueMakeNull
Creates a JavaScript value of the null type.

JSValueMakeNumber
Creates a JavaScript value of the number type.

JSValueMakeString
Creates a JavaScript value of the string type.

JSValueMakeUndefined
Creates a JavaScript value of the undefined type.

JSValueProtect
Protects a JavaScript value from garbage collection.

JSValueToBoolean
Converts a JavaScript value to boolean and returns the resulting boolean.

JSValueToNumber
Converts a JavaScript value to number and returns the resulting number.

JSValueToObject
Converts a JavaScript value to object and returns the resulting object.

JSValueToStringCopy
Converts a JavaScript value to string and copies the result into a JavaScript string.

JSValueUnprotect
Unprotects a JavaScript value from garbage collection.

Type Definitions

JSChar
A UTF-16 code unit.

JSClassAttributes

JSClassRef

JSContextGroupRef

JSContextRef

JSGlobalContextRef

JSObjectCallAsConstructorCallback

JSObjectCallAsFunctionCallback

JSObjectConvertToTypeCallback

JSObjectDeletePropertyCallback

JSObjectFinalizeCallback

JSObjectGetPropertyCallback

JSObjectGetPropertyNamesCallback

JSObjectHasInstanceCallback

JSObjectHasPropertyCallback

JSObjectInitializeCallback

JSObjectRef

JSObjectSetPropertyCallback

JSPropertyAttributes

JSPropertyNameAccumulatorRef

JSPropertyNameArrayRef

JSStringRef

JSTypedArrayBytesDeallocator

JSValueRef