Function javascriptcore_sys::JSCheckScriptSyntax

pub unsafe extern fn JSCheckScriptSyntax(ctx: JSContextRef,
                                         script: JSStringRef,
                                         sourceURL: JSStringRef,
                                         startingLineNumber: c_int,
                                         exception: *mut JSValueRef)
                                         -> bool

Checks for syntax errors in a string of JavaScript.

ctx: The execution context to use.

script: A JSString containing the script to check for syntax errors.

sourceURL: A JSString containing a URL for the script's source file. This is only used when reporting exceptions. Pass NULL if you do not care to include source file information in exceptions.

startingLineNumber: An integer value specifying the script's starting line number in the file located at sourceURL. This is only used when reporting exceptions. The value is one-based, so the first line is line 1 and invalid values are clamped to 1.

exception: A pointer to a JSValueRef in which to store a syntax error exception, if any. Pass NULL if you do not care to store a syntax error exception.

Returns true if the script is syntactically correct, otherwise false.