Function javascriptcore_sys::JSObjectMakeFunctionWithCallback

pub unsafe extern "C" fn JSObjectMakeFunctionWithCallback(
    ctx: JSContextRef, 
    name: JSStringRef, 
    callAsFunction: JSObjectCallAsFunctionCallback
) -> JSObjectRef

Convenience method for creating a JavaScript function with a given callback as its implementation.

ctx: The execution context to use.

name: A JSString containing the function's name. This will be used when converting the function to string. Pass NULL to create an anonymous function.

callAsFunction: The JSObjectCallAsFunctionCallback to invoke when the function is called.

Returns a JSObject that is a function. The object's prototype will be the default function prototype.