Function javascriptcore_sys::JSObjectSetPropertyAtIndex

pub unsafe extern fn JSObjectSetPropertyAtIndex(
    ctx: JSContextRef, 
    object: JSObjectRef, 
    propertyIndex: c_uint, 
    value: JSValueRef, 
    exception: *mut JSValueRef
)

Sets a property on an object by numeric index.

ctx: The execution context to use.

object: The JSObject whose property you want to set.

propertyIndex: The property's name as a number.

value: A JSValue to use as the property's value.

exception: A pointer to a JSValueRef in which to store an exception, if any. Pass NULL if you do not care to store an exception.

Calling JSObjectSetPropertyAtIndex is equivalent to calling JSObjectSetProperty with a string containing propertyIndex, but JSObjectSetPropertyAtIndex provides optimized access to numeric properties.