Function javascriptcore_sys::JSStringGetUTF8CString

pub unsafe extern fn JSStringGetUTF8CString(
    string: JSStringRef, 
    buffer: *mut c_char, 
    bufferSize: usize
) -> usize

Converts a JavaScript string into a null-terminated UTF8 string, and copies the result into an external byte buffer.

string: The source JSString.

buffer: The destination byte buffer into which to copy a null-terminated UTF8 representation of string. On return, buffer contains a UTF8 string representation of string. If bufferSize is too small, buffer will contain only partial results. If buffer is not at least bufferSize bytes in size, behavior is undefined.

bufferSize: The size of the external buffer in bytes.

Returns the number of bytes written into buffer (including the null-terminator byte).