Type Definition javascriptcore_sys::JSObjectSetPropertyCallback

type JSObjectSetPropertyCallback = Option<unsafe extern fn(_: JSContextRef, _: JSObjectRef, _: JSStringRef, _: JSValueRef, _: *mut JSValueRef) -> bool>;

The callback invoked when setting a property's value.

ctx: The execution context to use.

object: The JSObject on which to set the property's value.

propertyName: A JSString containing the name of the property to set.

value: A JSValue to use as the property's value.

exception: A pointer to a JSValueRef in which to return an exception, if any.

Returns true if the property was set, otherwise false.

If you named your function SetProperty, you would declare it like this:

bool
SetProperty(JSContextRef ctx, JSObjectRef object,
            JSStringRef propertyName, JSValueRef value,
            JSValueRef* exception);

If this function returns false, the set request forwards to object's statically declared properties, then its parent class chain (which includes the default object class).